[Chore #163083232] Add base class for models
